gros refactoring sur les méthodes pour invoquer une tour: on peut maintenant ajouter une tour grace aux symboles que l'on ajoute au pentacle et la posé sur le terrain grace a la souris

+++ src/main/java/universite_paris8/iut/nchaieb/sae_jeux/modele/Tour.java
@@ -4,7 +4,7 @@ import javafx.collections.ObservableList;
 
 public class Tour extends Entite{
     protected int portee;
-    protected int x, y;
+    protected double x, y;
 
     public Tour(int portee, int atq, int x, int y) {
